Major update * Fixed a bug in PMM * Implemented function to retrieve cpuinfo * Implemented a driver for keyboard * Implemented a new shell * Removed debug symbol support * Added a new data structure 'list' * Implemented a new Pseudorandom number generator * Added new string operators * Updated link.ld * Fixed all warnings except one in serial.c * Added a function to read timestamp counter * Added a function to measure cpu clock speed * Updated textmode driver * Updated Makefile * Incremented kernel version * Implemented a function to dump contents of a memory address to screen or serial port

+++ src/include/vga_text.h
@@ -26,8 +26,8 @@
 
 typedef struct vga_text
 {
-    uint32_t x;
-    uint32_t y;
+    uint16_t x;
+    uint16_t y;
 
     uint32_t height;
     uint32_t width;
